Remove axios and yargs from deploy package
