+++ typescript/deploy/package.json
@@ -8,10 +8,8 @@
     "@abacus-network/sdk": "^0.2.3",
     "@abacus-network/utils": "^0.2.3",
     "@types/debug": "^4.1.7",
-    "@types/yargs": "^17.0.10",
-    "axios": "^0.21.3",
-    "debug": "^4.3.4",
-    "yargs": "^17.4.1"
+    "cross-fetch": "^3.1.5",
+    "debug": "^4.3.4"
   },
   "devDependencies": {
     "@types/node": "^16.9.1",
